# schedule prefills considering padded shape