@tae-su-kim 2024-10-18
Closing this issue as #394 implements requested feature.